put adam and adamax together
